#1125 moved trimJSONWhitespace to validation functions, to catch whitespace before lang validations
