+++ src/controller/cve.controller/cve.controller.js
@@ -11,27 +11,6 @@ function createProviderMetadata (orgId, shortName, updateDate) {
   return { orgId: orgId, shortName: shortName, dateUpdated: updateDate }
 }
 
-// Helper function to trim trailing and leading whitespace from JSON field values
-function trimWhitespace (jsonBody) {
-  const queue = [jsonBody] // A queue to track objects for processing
-
-  while (queue.length > 0) {
-    const currentObject = queue.shift()
-
-    for (const key in currentObject) {
-      const value = currentObject[key]
-
-      if (typeof value === 'string') {
-        currentObject[key] = value.trim()
-      } else if (typeof value === 'object') {
-        queue.push(value) // Add nested objects to the queue
-      }
-    }
-  }
-
-  return jsonBody
-}
-
 // Called by GET /cve/:id
 async function getCve (req, res, next) {
   try {
@@ -352,9 +331,7 @@ async function submitCve (req, res, next) {
   const CONSTANTS = getConstants()
 
   try {
-    // Trim any leading/trailing whitespace from values
-    const trimmedJSON = trimWhitespace(req.ctx.body)
-    const newCve = new Cve({ cve: trimmedJSON })
+    const newCve = new Cve({ cve: req.ctx.body })
     const id = req.ctx.params.id
     const cveId = newCve.cve.cveMetadata.cveId
     const state = newCve.cve.cveMetadata.state
@@ -414,9 +391,7 @@ async function updateCve (req, res, next) {
   const CONSTANTS = getConstants()
 
   try {
-    // Trim any leading/trailing whitespace from values
-    const trimmedJSON = trimWhitespace(req.ctx.body)
-    const newCve = new Cve({ cve: trimmedJSON })
+    const newCve = new Cve({ cve: req.ctx.body })
     const cveId = req.ctx.params.id
     const cveRepo = req.ctx.repositories.getCveRepository()
     const cveIdRepo = req.ctx.repositories.getCveIdRepository()
@@ -506,9 +481,7 @@ async function submitCna (req, res, next) {
     const owningCna = await orgRepo.findOneByUUID(cveId.owning_cna)
     const assignerShortName = owningCna.short_name
 
-    // Trim any leading/trailing whitespace from values
-    const trimmedJSON = trimWhitespace(req.ctx.body)
-    const cnaContainer = trimmedJSON.cnaContainer
+    const cnaContainer = req.ctx.body.cnaContainer
     const dateUpdated = (new Date()).toISOString()
     const additionalCveMetadataFields = {
       assignerShortName: assignerShortName,
@@ -590,9 +563,7 @@ async function updateCna (req, res, next) {
 
     // update cve record here
     const cveRecord = result.cve
-    // Trim any leading/trailing whitespace from values
-    const trimmedJSON = trimWhitespace(req.ctx.body)
-    const cnaContainer = trimmedJSON.cnaContainer
+    const cnaContainer = req.ctx.body.cnaContainer
     const dateUpdated = (new Date()).toISOString()
     cveRecord.cveMetadata.dateUpdated = dateUpdated
