+++ src/controller/cve.controller/cve.middleware.js
@@ -130,7 +130,8 @@ function validateRejectBody (req, res, next) {
 }
 
 function validateCveCnaContainerJsonSchema (req, res, next) {
-  const cnaContainer = req.body
+  // Trims leading/trailing whitespace from JSON values
+  const cnaContainer = utils.trimJSONWhitespace(req.body)
   const result = validateCnaContainer(cnaContainer)
   if (!result) {
     logger.error(JSON.stringify({ uuid: req.ctx.uuid, message: 'CVE JSON schema validation FAILED.' }))
