+++ src/middleware/middleware.js
@@ -13,6 +13,7 @@ const errors = require('./error')
 const error = new errors.MiddlewareError()
 const RepositoryFactory = require('../repositories/repositoryFactory')
 const rateLimit = require('express-rate-limit')
+const trimJSONWhitespace = require('../utils/utils').trimJSONWhitespace
 
 function setCacheControl (req, res, next) {
   res.set('Cache-Control', 'no-store')
@@ -310,7 +311,9 @@ async function cnaMustOwnID (req, res, next) {
 
 function validateCveJsonSchema (req, res, next) {
   const CONSTANTS = getConstants()
-  const cve = req.body
+
+  // Trim any leading/trailing whitespace from values
+  const cve = trimJSONWhitespace(req.body)
   let cveState = cve.cveMetadata
 
   if (!cve.dataVersion) {
