+++ src/utils/utils.js
@@ -153,6 +153,27 @@ function toDate (val) {
   return result
 }
 
+// Helper function to trim trailing and leading whitespace from JSON field values
+function trimJSONWhitespace (jsonBody) {
+  const queue = [jsonBody] // A queue to track objects for processing
+
+  while (queue.length > 0) {
+    const currentObject = queue.shift()
+
+    for (const key in currentObject) {
+      const value = currentObject[key]
+
+      if (typeof value === 'string') {
+        currentObject[key] = value.trim()
+      } else if (typeof value === 'object') {
+        queue.push(value) // Add nested objects to the queue
+      }
+    }
+  }
+
+  return jsonBody
+}
+
 module.exports = {
   isSecretariat,
   isBulkDownload,
@@ -163,5 +184,6 @@ module.exports = {
   getUserUUID,
   reqCtxMapping,
   booleanIsTrue,
-  toDate
+  toDate,
+  trimJSONWhitespace
 }
